# BackEnd/swagger.yaml
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
# Added by API Auto Mocking Plugin
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/MatteoLunardon/TRENTunOAPIs_la_vendetta/1.0.0
- url: https://virtserver.swaggerhub.com/JOAOPROGRAMMING/TRENTunO/1.0.0
  description: TRENTunO API Auto Mocking
...
paths:
  /users:
    get:
      tags:
      - User
      description: This API will return all users in the system
      parameters:
      - name: user_level
        in: query
        description: Auth level of the user
        required: false
        schema:
          type: string
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                    $ref: '#/components/schemas/Users'
        "400":
          description: Invalid input format
        "404":
          description: User not found
        "500":
          description: Server Error
    post:
      tags:
      - User
      description: This API will add a new user to the system
      requestBody:
        description: The User object to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Users'
      responses:
        "201":
          description: successful operation
        "409":
          description: an existing item already exists
        "500":
          description: Server error
  /users/{id}:
    get:
      tags:
      - User
      description: This API will return the user in the system who has the id specified
      parameters:
        - name: id
          in: path
          description: User id
          required: true 
          schema: 
            $ref: '#/components/schemas/Users'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Users'
        "400":
          description: bad input parameter
    delete:
      tags:
      - User
      description: This API will delete the user in the system who has the id specified
      parameters:
        - name: id
          in: path
          description: User id
          required: true 
          schema:
              $ref: '#/components/schemas/Users'
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
                schema:
                    $ref: '#/components/schemas/Users'
              
        "400":
          description: Invalid user ID format
        "404":
          description: User not found
        "500":
          description: Server Error
    put:
      tags:
      - User
      description: This API will modify the values of the user in the system who has the id specified
      parameters:
      - name: id
        in: path
        description: User id
        required: true 
        schema:
          type: string
      requestBody:
        required: true
        schema:
          $ref: '#/components/schemas/Users'
      responses:
        "200":
          description: Successful operation
        "400":
          description: No fields to update provided
        "404":
          description: No user found with the provided ID
        "500":
          description: Server error
  /users/authentications:
    post:
      tags:
        - Users
      summary: Authenticate a user and return a JWT token
      description: Verifies email and password, and returns a JWT token for authenticated sessions.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "user@example.com"
                password:
                  type: string
                  format: password
                  example: "mypassword123"
      responses:
        "200":
          description: Authentication successful, token provided
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Authentication success"
                  token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  userId:
                    type: string
                    example: "65f8b2a456eaf23c0d123456"
                  auth:
                    type: boolean
                    example: true
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  message:
                    type: string
                    example: "Authentication failed"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Internal server error"
...
components:
  schemas:
    Users:
      type: object
      required:
        - name
        - surname
        - email
        - password
        - user_level
      properties:
        id:
          type: string
          description: Unique identifier for the user
          example: "60d21b4567d0d8992e610b98"
        name:
          type: string
          description: The user's first name
          example: "John"
        surname:
          type: string
          description: The user's last name
          example: "Doe"
        email:
          type: string
          description: The user's email address
          format: email
          example: "john.doe@example.com"
        password:
          type: string
          description: The user's password
          example: "$2b$10$N9qo8uLOickgx"
        user_level:
          type: string
          description: The user's level of access
          enum:
            - Admin
            - Client
          example: "Admin"
        auth:
          type: boolean
          description: Whether the user is authenticated
          default: false
          example: false
...